requirements.txt
watchdog
pyyaml
termcolor
cython
pygments
setproctitle
pylru
requests
redis
fastapi
uvicorn
dmoj